feat: 연습일정 타입 추가
